Change 'price' format in ProductDto to BigDecimal

+++ src/main/java/com/kodilla/ecommercee/domain/ProductDto.java
@@ -1,13 +1,15 @@
 package com.kodilla.ecommercee.domain;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private String groupId;
 
-    public ProductDto(Long id, String name, String description, Double price, String groupId) {
+    public ProductDto(Long id, String name, String description, BigDecimal price, String groupId) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -23,7 +25,7 @@ public class ProductDto {
         return name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
